DuyAkaJayD-Branch-BranchStatus

+++ src/main/java/org/edu/restaurantapi/service/BranchService.java
@@ -3,6 +3,9 @@ package org.edu.restaurantapi.service;
 import org.edu.restaurantapi.model.Branch;
 import org.edu.restaurantapi.repository.BranchRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
+import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -19,35 +22,54 @@ public class BranchService {
     }
 
     // Thêm
+    @PreAuthorize("hasRole('ADMIN')")
     public Branch createBranch(Branch branch) {
+        Optional<Branch> existingDiscount = branchRepository.findByPhoneNumber(branch.getPhoneNumber());
+        if (existingDiscount.isPresent()) {
+            return null;  // Trả về null nếu mã giảm giá đã tồn tại
+        }
         return branchRepository.save(branch);
     }
 
+
+
     // Xem
-    public List<Branch> getAllBranches() {
-        return branchRepository.findAll();
+    @PreAuthorize("hasRole('ADMIN')")
+    public Page<Branch> getAllBranches(Pageable pageable) {
+        return branchRepository.findAll(pageable);
     }
 
 // Tìm
-    public Optional<Branch> getBranchById(Long id) {
-        return branchRepository.findById(id);
+    public Branch getBranchById(Long id) {
+        return branchRepository.findById(id).orElse(null);
     }
 
     // Cập nhật
+    @PreAuthorize("hasRole('ADMIN')")
     public Branch updateBranch(Long id, Branch branchDetails) {
-        return branchRepository.findById(id).map(branch -> {
-            branch.setName(branchDetails.getName());
-            branch.setPhoneNumber(branchDetails.getPhoneNumber());
-            branch.setAddress(branchDetails.getAddress());
-            branch.setDistrict(branchDetails.getDistrict());
-            branch.setCity(branchDetails.getCity());
-            branch.setBranchStatus(branchDetails.getBranchStatus());
-            return branchRepository.save(branch);
+        return branchRepository.findById(id).map(existingBranch -> {
+            existingBranch.setName(branchDetails.getName() != null ? branchDetails.getName() : existingBranch.getName());
+            existingBranch.setPhoneNumber(branchDetails.getPhoneNumber() != null ? branchDetails.getPhoneNumber() : existingBranch.getPhoneNumber());
+            existingBranch.setAddress(branchDetails.getAddress() != null ? branchDetails.getAddress() : existingBranch.getAddress());
+            existingBranch.setDistrict(branchDetails.getDistrict() != null ? branchDetails.getDistrict() : existingBranch.getDistrict());
+            existingBranch.setCity(branchDetails.getCity() != null ? branchDetails.getCity() : existingBranch.getCity());
+            existingBranch.setBranchStatus(branchDetails.getBranchStatus() != null ? branchDetails.getBranchStatus() : existingBranch.getBranchStatus());
+            return branchRepository.save(existingBranch);
         }).orElseThrow(() -> new RuntimeException("Không tìm thấy id này " + id));
     }
 
+
     // Xóa
-    public void deleteBranch(Long id) {
-        branchRepository.deleteById(id);
+    @PreAuthorize("hasRole('ADMIN')")
+    public boolean deleteBranch(Long id) {
+        if (branchRepository.existsById(id)) {
+            branchRepository.deleteById(id);
+            return true;
+        }
+        return false;
+    }
+    //check
+    public boolean branchExists(Branch branch) {
+        return branchRepository.findByPhoneNumber(branch.getPhoneNumber()).isPresent();
     }
 }
